trying to expand to 32x8 from 16x8, incomplete

+++ 74595_shift.c
@@ -21,7 +21,7 @@
 		*(latch->port) &= ~(1<<(latch->pin_c));
         }
 }*/
-void shift_data(uint16_t data,uint8_t len,SHIFT_595_CB  *latch)
+void shift_data(uint32_t data,uint8_t len,SHIFT_595_CB  *latch)
 {
 
         uint8_t  i=0;
