+++ led_matrix.h
@@ -11,7 +11,7 @@ extern volatile uint8_t scroll_time;
 extern volatile uint8_t scanned_row;
 extern volatile uint8_t scroll;
 extern volatile uint16_t count;
-extern volatile uint16_t  display_rows[MATRIX_ROW_COUNT];
+extern volatile uint32_t  display_rows[MATRIX_ROW_COUNT];
 extern volatile uint8_t len;
 extern volatile uint8_t col_count;
 extern volatile uint8_t scroll;
@@ -22,6 +22,6 @@ extern SHIFT_595_CB latch;
 extern volatile char message[MAX_MESSAGE_LEN];
 extern char message_bak[MAX_MESSAGE_LEN];
 
-void scan_row(uint16_t byte,SHIFT_595_CB *latch);
+void scan_row(uint32_t byte,SHIFT_595_CB *latch);
 
 #endif
